+++ examples/dts-generation/src/pages/Home/index.tsx
@@ -1,0 +1,13 @@
+import store from './store';
+
+function Home() {
+  const [state] = store.useModel('counter');
+  return (
+    <div>
+      Home Page
+      <div>count: {state.count}</div>
+    </div>
+  );
+}
+
+export default Home;