+++ packages/build-user-config/src/userConfig/sassLoaderOptions.js
@@ -8,12 +8,26 @@ module.exports = (config, customOptions, context) => {
     defaultSassImplPkg = 'node-sass';
   }
   sassLoaderOptions.implementation = getSassImplementation(defaultSassImplPkg, rootDir);
+
+  // To prevent minimize .scss by default. And let css-minimizer-webpack-plugin take in control.
+  sassLoaderOptions.sassOptions = {
+    outputStyle: 'expanded'
+  };
+
   if (customOptions) {
     // the prependData option was removed in favor the additionalData option in sass-loader^10.0.0
     if (customOptions.prependData) {
       sassLoaderOptions.additionalData = customOptions.prependData;
       delete customOptions.prependData;
     }
+
+    if (customOptions.sassOptions) {
+      sassLoaderOptions.sassOptions = {
+        ...sassLoaderOptions.sassOptions,
+        ...customOptions.sassOptions
+      };
+    }
+
     sassLoaderOptions = { ...sassLoaderOptions, ...customOptions };
   }
   [