+++ packages/plugin-mpa/src/index.ts
@@ -10,9 +10,7 @@ interface ITemplate {
 interface IMpaConfig {
   template?: ITemplate;
   openPage?: string;
-  rewrites?: {
-    [key: string]: string;
-  };
+  rewrites?: {[key: string]: string} | boolean;
 }
 
 const plugin: IPlugin = (api) => {
@@ -88,9 +86,9 @@ const plugin: IPlugin = (api) => {
     let finalMPAEntries = {};
     if (parsedEntries) {
       Object.keys(parsedEntries).forEach((entryKey) => {
-        const { finalEntry, shouldRedirectRunApp, runAppPath } = parsedEntries[entryKey];
+        const { finalEntry, runAppPath } = parsedEntries[entryKey];
         finalMPAEntries[entryKey] = finalEntry;
-        if (shouldRedirectRunApp) {
+        if (runAppPath) {
           redirectEntries.push({
             entryPath: finalEntry,
             runAppPath,
@@ -114,15 +112,17 @@ const plugin: IPlugin = (api) => {
     // set page template
     onGetWebpackConfig(config => {
       setPageTemplate(rootDir, entries, (mpa as any).template || {}, config, setValue);
-
-      config.devServer.historyApiFallback({
-        rewrites: Object.keys(entries).map((pageName) => {
-          return {
-            from: new RegExp(`^/${mpaRewrites[pageName] || pageName}/*`),
-            to: `/${pageName}.html`,
-          };
-        })
-      });
+      // disable mpa rewrite rules when config mpa.rewrites as false
+      if ((userConfig.mpa as IMpaConfig)?.rewrites !== false) {
+        config.devServer.historyApiFallback({
+          rewrites: Object.keys(entries).map((pageName) => {
+            return {
+              from: new RegExp(`^/${mpaRewrites[pageName] || pageName}/*`),
+              to: `/${pageName}.html`,
+            };
+          })
+        });
+      }
     });
   }
 };