+++ packages/vite-service/src/wp2vite/config.ts
@@ -93,7 +93,8 @@ const configMap: ConfigMap = {
     name: 'resolve.alias',
     transform: (value, ctx) => {
       const { rootDir } = ctx;
-      const blackList = ['webpack/hot', 'node_modules'];
+      // webpack/hot is not necessary in mode vite
+      const blackList = ['webpack/hot'];
       const data: Record<string, any> = Object.keys(value).reduce(
         (acc, key) => {
           if (!blackList.some((word) => value[key]?.includes(word)))
@@ -158,13 +159,22 @@ const configMap: ConfigMap = {
   'devServer.proxy': {
     name: 'server.proxy',
     transform: (value: Record<string, any>[]) => {
-      // vite proxy do not support config of onProxyRes, onError, logLevel
+      // vite proxy do not support config of onProxyRes, onError, logLevel, pathRewrite
       // transform devServer.proxy to server.proxy
       const proxyConfig = {};
-      (value || []).forEach(({ context, enable, onProxyRes, onError, ...rest }) => {
+      (value || []).forEach(({ context, enable, onProxyRes, onError, pathRewrite, ...rest }) => {
         if (enable !== false) {
           proxyConfig[context] = {
             ...rest,
+            rewrite: (requestPath: string) => {
+              if (pathRewrite && isObject(pathRewrite)) {
+                return Object.keys(pathRewrite).reduce((acc, rewriteRule) => {
+                  return acc.replace(new RegExp(rewriteRule), pathRewrite[rewriteRule]);
+                }, requestPath);
+              } else {
+                return requestPath;
+              }
+            },
             configure: (proxy, options) => {
               if (rest.configure) {
                 rest.configure(proxy, options);
@@ -217,11 +227,12 @@ const configMap: ConfigMap = {
       if (userConfig?.postcssOptions) {
         const postcssPlugins = (userConfig?.postcssOptions as PostcssOptions)?.plugins || {};
         const normalizedPlugins = Object.keys(postcssPlugins)
-          .filter((pluginKey) => !postcssPlugins[pluginKey])
-          .map(pluginKey => [pluginKey, postcssPlugins[pluginKey]]);
+          .filter((pluginKey) => !!postcssPlugins[pluginKey])
+          // eslint-disable-next-line global-require,import/no-dynamic-require
+          .map(pluginKey => require(pluginKey)(postcssPlugins[pluginKey]));
         return {
           ...(userConfig?.postcssOptions as PostcssOptions),
-          plugins: normalizedPlugins.length > 1 ? normalizedPlugins : [],
+          plugins: normalizedPlugins.length > 0 ? normalizedPlugins : [],
         };
       }
     }