+++ packages/build-webpack-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+## 1.0.3
+
+- [feat] add postcss-nested for css nesting rules
+
 ## 1.0.2
 
 - [fix] bump version of `@builder/pack`